+++ cmd/opentelemetry/app/internal/esutil/index_name_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package esclient
+package esutil
 
 import (
 	"testing"
